[ticket/15167] Revert bitfield changes
PHPBB3-15167
