+++ phpBB/phpbb/feed/topic.php
@@ -78,7 +78,7 @@ class topic extends post_base
 		}
 
 		// Make sure forum is not excluded from feed
-		if (\phpbb\bitfield::optionget(FORUM_OPTION_FEED_EXCLUDE, $this->topic_data['forum_options']))
+		if (phpbb_optionget(FORUM_OPTION_FEED_EXCLUDE, $this->topic_data['forum_options']))
 		{
 			throw new no_feed_exception();
 		}
