+++ tests/bootstrap.php
@@ -23,7 +23,6 @@ require_once $phpbb_root_path . 'includes/startup.php';
 $table_prefix = 'phpbb_';
 require_once $phpbb_root_path . 'includes/constants.php';
 require_once $phpbb_root_path . 'phpbb/class_loader.' . $phpEx;
-require_once $phpbb_root_path . 'phpbb/bitfield.' . $phpEx;
 require_once $phpbb_root_path . 'includes/utf/utf_tools.' . $phpEx;
 require_once $phpbb_root_path . 'includes/functions.' . $phpEx;
 require_once $phpbb_root_path . 'includes/functions_content.' . $phpEx;
